+++ testdata/pad/transforms/times_by_day_of_week.sql
@@ -1,0 +1,1 @@
+SELECT day_of_week, COUNT(*) FROM public.worldtimeapi_ip_times GROUP BY 1;